[CV-423] Refactored useFetchData.ts and nativePropertyManagementScreen.tsx, placed duplicated code as a function into processCondoUnits. Reduces duplicability
